move configs from common to config
