+++ lib/constructPlugin.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var config = require('../common/config');
+var config = require('../config');
 
 var getPlugin = function getPlugin(name) {
   return config.plugins.filter(function (plugin) {
