+++ lib/events.js
@@ -4,7 +4,7 @@ var moment = require('moment');
 var parser = require('markdown-parse');
 var S = require('string');
 var GitHubApi = require('github');
-var config = require('../common/config');
+var config = require('../config');
 var milestone = require('./milestone');
 
 var createOrUpdateEvent = function(payload, event, sender, github) {
