+++ lib/milestone.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var debug = require('debug')('gitevents-milestone');
-var config = require('../common/config');
+var config = require('../config');
 
 var createMilestone = function createMilestone(payload, event, repo, github) {
   return new Promise(function(resolve, reject) {
