+++ lib/talks.js
@@ -3,7 +3,7 @@ var parser = require('markdown-parse');
 var moment = require('moment');
 var S = require('string');
 var GitHubApi = require('github');
-var config = require('../common/config');
+var config = require('../config');
 
 function checkMilestone(milestone, github, args) {
   return new Promise(function(resolve, reject) {
