+++ nodemon.json
@@ -1,4 +1,4 @@
 {
   "verbose": true,
-  "watch": ["lib/**/*", "common/**/*", "gitevents.js"]
+  "watch": ["lib/**/*", "config/**/*", "gitevents.js"]
 }
